Add new version of CDTest from DBIx::OnlineDDL
Should this be its own CPAN distro?  Maybe, but given the drastic changes to
its API so far, I don't think it's quite ready for public use yet.  Modules
like DBIx::Class::{Sims,Fixtures} and Test::DBIx::Class are also out there, so
it's possible we may decide on using one of those in the future.