Enable fractional multipliers in chunk resizing
Using chunk resizing with non-unique IDs presents a challenge for cases where
the first chunk ends up being larger than an acceptable chunk size.  The
previously integered multiplier goes down from 1 to 0, causing an infinite
loop that breaks the whole process.

Move the int function to a ceil on the end ID logic, and now the ID references
are still protected as integers.