Use ~0 as the max integer for $DB_MAX_ID
The previously-used Data::Float::max_integer value uses a NV (big surprise)
to represent this, which generally stringifies into an unsafe and unwanted
scientific notation string for values larger than a cerain size.  The ~0
trick, OTOH, will use a UV, which is always represented in the correct format.

As an added bonus, I no longer need the module as a dependency.